Separate `StringError` and `errorf` from  `parser`
