+++ kicad_rs/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod codec;
+pub mod error;
 pub mod parser;
 pub mod resolver;
 pub mod types;
